perf(emoji): move the EMOJI_SLUGS list out of the player bundle
The 128-entry slug list lived in ui/emoji.js, which player views
(draw, collection) import for emojiUrl / emojiImgHTML. ES modules
don't tree-shake, so every player parsed ~1 KB of data they never
touched. Move EMOJI_SLUGS to features/admin/emoji-slugs.js (only
admin/cards.js imports it, for the datalist autocomplete) and add
the new file to the service worker shell list so it stays available
offline for admins.
